Rework session managment and make multiplayer toggle-able.
Multiplayer is enabled if 'multiplayer.txt' exists in the
storage directory.

+++ zwift_offline.py
@@ -11,13 +11,19 @@ import sys
 import tempfile
 import time
 import math
-import threading
 from copy import copy
 from datetime import timedelta
+from functools import wraps
 from io import BytesIO
 from shutil import copyfile
+if sys.version_info[0] > 2:
+    from urllib.parse import quote, unquote
+else:
+    from urllib import quote, unquote
 
+import jwt
 from flask import Flask, request, jsonify, g, redirect, render_template, url_for, flash, session, abort
+from flask_login import UserMixin, AnonymousUserMixin, LoginManager, login_user, current_user, login_required
 from google.protobuf.descriptor import FieldDescriptor
 from protobuf_to_dict import protobuf_to_dict, TYPE_CALLABLE_MAP
 from flask_sqlalchemy import sqlalchemy, SQLAlchemy
@@ -70,6 +76,9 @@ DATABASE_CUR_VER = 2
 # For auth server
 AUTOLAUNCH_FILE = "%s/auto_launch.txt" % STORAGE_DIR
 SERVER_IP_FILE = "%s/server-ip.txt" % STORAGE_DIR
+MULTIPLAYER = False
+if os.path.exists("%s/multiplayer.txt" % STORAGE_DIR):
+    MULTIPLAYER = True
 from tokens import *
 
 AUTH_PATH = "%s/auth.db" % STORAGE_DIR
@@ -87,20 +96,30 @@ ghostsEnabled = {}
 playerUpdateQueue = {}
 playerIds = {}
 playerPartialProfiles = {}
-temporaryLoginValues = {}
 saveGhost = None
 
-class User(db.Model):
-    uid = db.Column(db.Integer, primary_key=True)
+class User(UserMixin, db.Model):
+    player_id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(100), unique=True, nullable=False)
     first_name = db.Column(db.String(100), nullable=False)
     last_name = db.Column(db.String(100), nullable=False)
     pass_hash = db.Column(db.String(100), nullable=False)
-    machine_id = db.Column(db.String(100), nullable=True)
     enable_ghosts = db.Column(db.Integer, nullable=False, default=1)
 
     def __repr__(self):
-        return '' % self.username
+        return self.username
+
+    def get_id(self):
+        return self.player_id
+
+class AnonUser(User, AnonymousUserMixin, db.Model):
+    username = "zoffline"
+    first_name = "z"
+    last_name = "offline"
+    enable_ghosts = True
+
+    def is_authenticated(self):
+        return True
 
 class PartialProfile:
     first_name = ''
@@ -153,33 +172,7 @@ def getOnline():
             onlineInRegion.paris += 1
         onlineInRegion.total += 1
     return onlineInRegion
-            
 
-def setMachineIdAndGhostsEnabled(request, player_id, ghosts_enabled):
-    machine_id = request.headers.get('X_MACHINE_ID')
-    #Remove machine id for other users
-    users = User.query.filter_by(machine_id=machine_id).all()
-    for user in users:
-        user.machine_id = None
-    #Set machine id for current user
-    user = User.query.filter_by(uid=player_id-1000).first()
-    user.machine_id = machine_id
-    user.enable_ghosts = int(ghosts_enabled)
-    db.session.commit()
-    playerIds[machine_id] = str(player_id)
-    ghostsEnabled[str(player_id)] = ghosts_enabled
-
-def getPlayerId(request):
-    machine_id = request.headers.get('X_MACHINE_ID')
-    ghosts_enabled = True
-    if not machine_id in playerIds:
-        user = User.query.filter_by(machine_id=machine_id).first()
-        playerIds[machine_id] = str(user.uid+1000)
-        ghosts_enabled = bool(user.enable_ghosts)
-    player_id = playerIds[machine_id]
-    if not player_id in ghostsEnabled:
-        ghostsEnabled[player_id] = ghosts_enabled
-    return player_id
 
 def getPartialProfile(player_id):
     player_id_str = str(player_id)
@@ -221,10 +214,26 @@ def isNearby(player_state1, player_state2, range = 100000):
                         return True
     except:
         pass
-
-
     return False
-                
+
+
+# We store flask-login's cookie in the "fake" JWT that we give Zwift.
+# Make it a cookie again to reuse flask-login on API calls.
+def jwt_to_session_cookie(f):
+    @wraps(f)
+    def wrapper(*args, **kwargs):
+        if not MULTIPLAYER:
+            return f(*args, **kwargs)
+        token = request.headers.get('Authorization')
+        if token and not session.get('_user_id'):
+            token = jwt.decode(token.split()[1], options=({'verify_signature': False, 'verify_aud': False}))
+            request.cookies = request.cookies.copy()  # request.cookies is an immutable dict
+            request.cookies['remember_token'] = token['session_cookie']
+            login_manager._load_user()
+
+        return f(*args, **kwargs)
+    return wrapper
+
 
 @app.route("/signup/", methods=["GET", "POST"])
 def signup():
@@ -268,9 +277,7 @@ def login():
         user = User.query.filter_by(username=username).first()
 
         if user and check_password_hash(user.pass_hash, password):
-            session[username] = True
-            temporaryLoginValues[request.remote_addr + ":player_id"] = user.uid + 1000
-            online = Online()
+            login_user(user, remember=True)
             return redirect(url_for("user_home", username=username, enable_ghosts=bool(user.enable_ghosts), online=getOnline()))
         else:
             flash("Invalid username or password.")
@@ -279,21 +286,15 @@ def login():
 
 
 @app.route("/user/<username>/")
+@login_required
 def user_home(username):
-    if not session.get(username):
-        abort(401)
-
-    user = User.query.filter_by(username=username).first()
-    online = Online()
-    return render_template("user_home.html", username=username, enable_ghosts=bool(user.enable_ghosts), online=getOnline())
+    return render_template("user_home.html", username=current_user.username, enable_ghosts=bool(current_user.enable_ghosts), online=getOnline())
 
 
 @app.route("/upload/<username>/", methods=["GET", "POST"])
+@login_required
 def upload(username):
-    if not session.get(username):
-        abort(401)
-
-    player_id = temporaryLoginValues[request.remote_addr + ":player_id"]
+    player_id = current_user.player_id
     profile_dir = os.path.join(STORAGE_DIR, str(player_id))
     try:
         if not os.path.isdir(profile_dir):
@@ -331,7 +332,6 @@ def upload(username):
 
 @app.route("/logout/<username>")
 def logout(username):
-    session.pop(username, None)
     flash("Successfully logged out.")
     return redirect(url_for('login'))
 
@@ -430,9 +430,11 @@ def api_users_login():
 
 
 @app.route('/api/users/logout', methods=['POST'])
+@jwt_to_session_cookie
+@login_required
 def api_users_logout():
     #Remove player from online when leaving game/world
-    player_id = getPlayerId(request)
+    player_id = current_user.player_id
     if player_id in online:
         online.pop(player_id)
     if player_id in playerPartialProfiles:
@@ -487,14 +489,26 @@ def api_telemetry_config():
 
 
 @app.route('/api/profiles/me', methods=['GET'])
+@jwt_to_session_cookie
+@login_required
 def api_profiles_me():
-    stored_player_id = temporaryLoginValues[request.remote_addr + ":player_id"]
-    stored_enable_ghosts = temporaryLoginValues[request.remote_addr + ":enableghosts"]
-    temporaryLoginValues.pop(request.remote_addr + ":player_id")
-    temporaryLoginValues.pop(request.remote_addr + ":enableghosts")
-    setMachineIdAndGhostsEnabled(request, stored_player_id, stored_enable_ghosts)
-    profile_id = int(stored_player_id)
-    profile_dir = '%s/%s' % (STORAGE_DIR, profile_id)
+    profile_id = current_user.player_id
+    if MULTIPLAYER:
+        profile_dir = '%s/%s' % (STORAGE_DIR, profile_id)
+    else:
+        # Find first profile.bin if one exists and use it. Multi-profile
+        # support is deprecated and now unsupported for non-multiplayer mode.
+        profile_dir = None
+        for name in os.listdir(STORAGE_DIR):
+            path = "%s/%s" % (STORAGE_DIR, name)
+            if os.path.isdir(path) and os.path.exists("%s/profile.bin" % path):
+                profile_dir = path
+                break
+        if not profile_dir:  # no existing profile
+            profile_dir = "%s/1" % STORAGE_DIR
+            profile_id = 1
+            AnonUser.player_id = profile_id
+
     try:
         if not os.path.isdir(profile_dir):
             os.makedirs(profile_dir)
@@ -506,43 +520,53 @@ def api_profiles_me():
     if not os.path.isfile(profile_file):
         profile.id = profile_id
         profile.is_connected_to_strava = True
-        user = User.query.filter_by(uid=(profile_id-1000)).first()
-        profile.email = user.username
-        profile.first_name = user.first_name
-        profile.last_name = user.last_name
+        profile.email = current_user.username
+        profile.first_name = current_user.first_name
+        profile.last_name = current_user.last_name
         return profile.SerializeToString(), 200
     with open(profile_file, 'rb') as fd:
         profile.ParseFromString(fd.read())
-        profile.id = profile_id
+        if MULTIPLAYER:
+            profile.id = profile_id
+        elif current_user.player_id != profile.id:
+            # Update AnonUser's player_id to match
+            AnonUser.player_id = profile.id
         if not profile.email:
             profile.email = 'user@email.com'
         if profile.f60:
-            # remove free trial limit
             del profile.f60[:]
         return profile.SerializeToString(), 200
 
 
 @app.route('/api/profiles/<int:player_id>', methods=['PUT'])
+@jwt_to_session_cookie
+@login_required
 def api_profiles_id(player_id):
     if not request.stream:
         return '', 400
+    if current_user.player_id != player_id:
+        return '', 401
     stream = request.stream.read()
     with open('%s/%s/profile.bin' % (STORAGE_DIR, player_id), 'wb') as f:
         f.write(stream)
     profile = profile_pb2.Profile()
     profile.ParseFromString(stream)
-    user = User.query.filter_by(uid=(player_id-1000)).first()
-    user.first_name = profile.first_name
-    user.last_name = profile.last_name
-    db.session.commit()
+    if MULTIPLAYER:
+        current_user.first_name = profile.first_name
+        current_user.last_name = profile.last_name
+        db.session.commit()
     return '', 204
 
 
 @app.route('/api/profiles/<int:player_id>/activities/', methods=['GET', 'POST'], strict_slashes=False)
+@jwt_to_session_cookie
+@login_required
 def api_profiles_activities(player_id):
     if request.method == 'POST':
         if not request.stream:
             return '', 400
+        if current_user.player_id != player_id:
+            return '', 401
         activity = activity_pb2.Activity()
         activity.ParseFromString(request.stream.read())
         activity.id = get_id('activity')
@@ -672,9 +696,13 @@ def garmin_upload(player_id, activity):
 # With 64 bit ids Zwift can pass negative numbers due to overflow, which the flask int
 # converter does not handle so it's a string argument
 @app.route('/api/profiles/<int:player_id>/activities/<string:activity_id>', methods=['PUT'])
+@jwt_to_session_cookie
+@login_required
 def api_profiles_activities_id(player_id, activity_id):
     if not request.stream:
         return '', 400
+    if current_user.player_id != player_id:
+        return '', 401
     activity_id = int(activity_id) & 0xffffffffffffffff
     activity = activity_pb2.Activity()
     activity.ParseFromString(request.stream.read())
@@ -683,8 +711,8 @@ def api_profiles_activities_id(player_id, activity_id):
     response = '{"id":%s}' % activity_id
     if request.args.get('upload-to-strava') != 'true':
         return response, 200
-    player_id = getPlayerId(request)
-    if ghostsEnabled.get(player_id):
+    player_id = current_user.player_id
+    if current_user.enable_ghosts:
         try:
             saveGhost(activity.name, int(player_id))
         except:
@@ -697,6 +725,8 @@ def api_profiles_activities_id(player_id, activity_id):
     return response, 200
 
 @app.route('/api/profiles/<int:recieving_player_id>/activities/0/rideon', methods=['POST']) #activity_id Seem to always be 0, even when giving ride on to ppl with 30km+
+@jwt_to_session_cookie
+@login_required
 def api_profiles_activities_rideon(recieving_player_id):
     sending_player_id = request.json['profileId']
     profile = getPartialProfile(sending_player_id)
@@ -799,7 +829,11 @@ def set_goal_end_date(goal, now):
 
 
 @app.route('/api/profiles/<int:player_id>/goals', methods=['GET', 'POST'])
+@jwt_to_session_cookie
+@login_required
 def api_profiles_goals(player_id):
+    if player_id != current_user.player_id:
+        return '', 401
     if request.method == 'POST':
         if not request.stream:
             return '', 400
@@ -833,7 +867,11 @@ def api_profiles_goals(player_id):
 
 
 @app.route('/api/profiles/<int:player_id>/goals/<string:goal_id>', methods=['DELETE'])
+@jwt_to_session_cookie
+@login_required
 def api_profiles_goals_id(player_id, goal_id):
+    if player_id != current_user.player_id:
+        return '', 401
     goal_id = int(goal_id) & 0xffffffffffffffff
     cur = g.db.cursor()
     cur.execute("DELETE FROM goal WHERE id = ?", (str(goal_id),))
@@ -959,7 +997,7 @@ def relay_worlds_generic(world_id=None):
             return world.SerializeToString()
         else:
             return worlds.SerializeToString()
-            
+
 
 @app.route('/relay/worlds', methods=['GET'])
 @app.route('/relay/dropin', methods=['GET'])
@@ -1099,9 +1137,13 @@ def relay_segment_results():
 
 
 @app.route('/api/segment-results', methods=['GET', 'POST'])
+@jwt_to_session_cookie
+@login_required
 def api_segment_results():
     #Checks that online player has values for ghosts and player_id
-    player_id = getPlayerId(request)
+    player_id = current_user.player_id
+    if request.method == 'POST' and player_id != current_user.player_id:
+        return '', 401
     return handle_segment_results(request)
 
 
@@ -1201,9 +1243,27 @@ def auth_rb():
 def launch_zwift():
     # Zwift client has switched to calling https://launcher.zwift.com/launcher/ride
     if request.path != "/ride" and not os.path.exists(AUTOLAUNCH_FILE):
-        return render_template("login_form.html")
+        if MULTIPLAYER:
+            return render_template("login_form.html")
+        else:
+            return render_template("user_home.html", username="", enable_ghosts=False, online=getOnline())
     else:
-        return redirect("http://zwift/?code=zwift_refresh_token%s" % REFRESH_TOKEN, 302)
+        if MULTIPLAYER:
+            return redirect("http://zwift/?code=%s" % quote(request.cookies.get('remember_token')), 302)
+        else:
+            return redirect("http://zwift/?code=zwift_refresh_token%s" % REFRESH_TOKEN, 302)
+
+
+def fake_jwt_with_session_cookie(session_cookie):
+    access_token = jwt.decode(ACCESS_TOKEN, options=({'verify_signature': False, 'verify_aud': False}))
+    access_token['session_cookie'] = session_cookie
+    access_token = jwt.encode(access_token, 'nosecret').decode('utf-8')
+
+    refresh_token = jwt.decode(REFRESH_TOKEN, options=({'verify_signature': False, 'verify_aud': False}))
+    refresh_token['session_cookie'] = session_cookie
+    refresh_token = jwt.encode(refresh_token, 'nosecret').decode('utf-8')
+
+    return """{"access_token":"%s","expires_in":1000021600,"refresh_expires_in":611975560,"refresh_token":"%s","token_type":"bearer","id_token":"%s","not-before-policy":1408478984,"session_state":"0846ab9a-765d-4c3f-a20c-6cac9e86e5f3","scope":""}""" % (access_token, refresh_token, ID_TOKEN)
 
 
 @app.route('/auth/realms/zwift/protocol/openid-connect/token', methods=['POST'])
@@ -1212,20 +1272,34 @@ def auth_realms_zwift_protocol_openid_connect_token():
     username = request.form.get('username')
     password = request.form.get('password')
 
-    if username:
+    if username and MULTIPLAYER:
         user = User.query.filter_by(username=username).first()
 
         if user and check_password_hash(user.pass_hash, password):
-            temporaryLoginValues[request.remote_addr + ":player_id"] = user.uid + 1000
+            login_user(user, remember=True)
         else:
             return '', 401
 
-    return FAKE_JWT, 200
-
+    if MULTIPLAYER:
+        # This is called once with ?code= in URL and once again with the refresh token
+        if "code" in request.form:
+            # Original code argument is replaced with session cookie from launcher
+            session_cookie = unquote(request.form['code'])
+            return fake_jwt_with_session_cookie(session_cookie), 200
+        elif "refresh_token" in request.form:
+            token = jwt.decode(request.form['refresh_token'], options=({'verify_signature': False, 'verify_aud': False}))
+            return fake_jwt_with_session_cookie(token['session_cookie'])
+    else:
+        return FAKE_JWT, 200
 
 @app.route("/start-zwift" , methods=['POST'])
 def start_zwift():
-    temporaryLoginValues[request.remote_addr + ":enableghosts"] = 'enableghosts' in request.form.keys()
+    if MULTIPLAYER:
+        current_user.enable_ghosts = 'enableghosts' in request.form.keys()
+        ghostsEnabled[str(current_user.player_id)] = current_user.enable_ghosts
+    else:
+        AnonUser.enable_ghosts = 'enableghosts' in request.form.keys()
+    db.session.commit()
     selected_map = request.form['map']
     if selected_map == 'CALENDAR':
         return redirect("/ride", 302)
@@ -1236,7 +1310,18 @@ def start_zwift():
 # Called by Mac, but not Windows
 @app.route('/auth/realms/zwift/tokens/access/codes', methods=['POST'])
 def auth_realms_zwift_tokens_access_codes():
-    return FAKE_JWT, 200
+    if MULTIPLAYER:
+        if "code" in request.form:
+            remember_token = unquote(request.form['code'])
+            return fake_jwt_with_session_cookie(remember_token), 200
+        elif "refresh_token" in request.form:
+            token = jwt.decode(request.form['refresh_token'], options=({'verify_signature': False, 'verify_aud': False}))
+            return fake_jwt_with_session_cookie(token['session_cookie'])
+        remember_token = unquote(request.form['code'])
+        return fake_jwt_with_session_cookie(remember_token), 200
+    else:
+        return FAKE_JWT, 200
+
 
 @app.route('/static/web/launcher/<filename>', methods=['GET'])
 def static_web_launcher(filename):
@@ -1270,12 +1355,24 @@ def run_standalone(passedOnline, passedGhostsEnabled, passedSaveGhost, passedPla
     global ghostsEnabled
     global saveGhost
     global playerUpdateQueue
+    global login_manager
     online = passedOnline
     ghostsEnabled = passedGhostsEnabled
     saveGhost = passedSaveGhost
     playerUpdateQueue = passedPlayerUpdateQueue
-    thread = threading.Thread(target=check_columns)
-    thread.start()
+    login_manager = LoginManager()
+    login_manager.login_view = 'login'
+    login_manager.session_protection = None
+    db.create_all(app=app)
+    db.session.commit()
+    if not MULTIPLAYER:
+        login_manager.anonymous_user = AnonUser
+    login_manager.init_app(app)
+
+    @login_manager.user_loader
+    def load_user(uid):
+        return User.query.get(int(uid))
+
     app.run(ssl_context=('%s/cert-zwift-com.pem' % SSL_DIR, '%s/key-zwift-com.pem' % SSL_DIR), port=443, threaded=True, host='0.0.0.0') # debug=True, use_reload=False)
